+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Ability to copy multiple blocks content by concatenating them in the order they were selected.
 - Added `Ctrl+E` and `Ctrl+P` note block shortcuts to switch between edit and preview modes while preserving inline code formatting and command palette conflicts.
+- Added `Duplicate` option in the block context menu to quickly duplicate a block.
 
 ### Fixed